+++ src/sit.cpp
@@ -911,7 +911,7 @@ Restart:
 		residuals();
 		for (j = 0; j < count_unknowns; j++)
 		{
-			my_array[j * (count_unknowns + 1) + i] =
+			my_array[(size_t)j * ((size_t)count_unknowns + 1) + (size_t)i] =
 				-(residual[j] - base[j]) / d2;
 		}
 		switch (x[i]->type)
@@ -930,9 +930,9 @@ Restart:
 			break;
 		case MH:
 			s_eminus->la -= d;
-			if (my_array[i * (count_unknowns + 1) + i] == 0)
+			if (my_array[(size_t)i * ((size_t)count_unknowns + 1) + (size_t)i] == 0)
 			{
-				my_array[i * (count_unknowns + 1) + i] =
+				my_array[(size_t)i * ((size_t)count_unknowns + 1) + (size_t)i] =
 					exp(s_h2->lm * LOG_10) * 2;
 			}
 			break;
